Initialize the solution graph and edge list
