+++ test/Makefile.am
@@ -12,7 +12,11 @@ AM_LDFLAGS = -module -rpath $(libdir) -avoid-version -no-undefined
 test_davenport_la_SOURCES = \
   test_helper.c \
   test_davenport.c \
+  test_dv_init_elist.c \
+  test_dv_init_solution.c \
+  test_edge_selection.c \
   test_lower_bound.c \
+  test_network.c \
   test_preflow_discharge.c \
   test_preflow_list_mgr.c \
   test_preflow_max_flow.c \
